third_party/WebKit/LayoutTests/shapedetection/detection-options.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/mock-facedetection.js"></script>
<body>
<img id='img' src='../media/content/greenbox.png'/>
</body>
<script>
async_test(function(t) {
  var img = document.getElementById("img");
  var theMock = null;
  mockFaceDetectionReady
    .then(mock => {
      theMock = mock;
      return new FaceDetector();
    })
    .catch(error => {
      assert_unreached("Error creating MockShapeDetection: " + error);
    })
    .then(detectorWithDefault => detectorWithDefault.detect(img))
    .then(t.step_func(faceDetectionResult => {
        assert_equals(theMock.getMaxDetectedFaces(), 10, "default maxDetectedFaces");
        assert_equals(theMock.getFastMode(), false, "default maxDetectedFaces");
        return new FaceDetector({maxDetectedFaces: 7, fastMode: true});
    }))
    .then(detectorWithOptions => detectorWithOptions.detect(img))
    .then(t.step_func(faceDetectionResult => {
        assert_equals(theMock.getMaxDetectedFaces(), 7, "maxDetectedFaces");
        assert_equals(theMock.getFastMode(), true, "maxDetectedFaces");
        t.done();
    }))
    .catch(error => {
      assert_unreached("Error creating detectors: " + error);
    });
}, "Test that FaceDetectionOptions are correctly propagated");
</script>
